refactor: [FE] 중복되는 getFetchData() fetchUtil 로 분리
